// src/app/types.ts
export interface City {
    lat: number
    lon: number
    country: string
    state: string
    name: string
    id: string
}

interface Weather {
    id: number
    main: string
    description: string
    icon: string
}

interface HornyWeather {
    dt: number
    temp: number
    feels_like: number
    pressure: number
    humidity: number
    dew_point: number
    uvi: number
    clouds: number
    visibility: number
    wind_speed: number
    wind_deg: number
    wind_gust: number
    weather: Weather[]
    pop: number
}

export interface DailyWeather {
    dt: number
    sunrise: number
    sunset: number
    moonrise: number
    moonset: number
    moon_phase: number
    temp: {
        day: number
        min: number
        max: number
        night: number
        eve: number
        morn: number
    }
    feels_like: {
        day: number
        night: number
        eve: number
        morn: number
    }
    pressure: number
    humidity: number
    dew_point: number
    wind_speed: number
    wind_deg: number
    wind_gust: number
    weather: Weather[]
    clouds: number
    pop: number
    rain: number
    uvi: number
}

export interface CurrentWeather {
    dt: number
    sunrise: number
    sunset: number
    temp: number
    feels_like: number
    pressure: number
    humidity: number
    dew_point: number
    uvi: number
    clouds: number
    visibility: number
    wind_speed: number
    wind_deg: number
    wind_gust: number
    weather: Weather[]
}

interface Response {
    lat: number
    lon: number
    timezone: string
    timezone_offset: number
    current: CurrentWeather
    hourly: HornyWeather[]
    daily: DailyWeather[]
    id: string
}

type Visibility = {
    visibility?: number
}

export type CurrentResponse = Response & Visibility